feat : adapter generation
