+++ hw_starship_battle/src/test/java/ru/otus/starshipbattletests/core/IoCExceptionTest.java
@@ -1,0 +1,14 @@
+package ru.otus.starshipbattletests.core;
+
+import org.junit.jupiter.api.Test;
+import ru.otus.starshipbattle.core.IoC;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class IoCExceptionTest {
+
+    @Test
+    void IocShouldThrowArgumentExceptionIfDependencyIsNotFound() {
+        assertThrows(RuntimeException.class, () -> IoC.resolve("IoC.Register"));
+    }
+}
